Visual Adjustments
Made Adjustments to the layout of both the SIR model and ZECC model so that graphs are easier to see on the website
